+++ apps/commerce-api/src/main/java/com/loopers/application/user/UserFacade.java
@@ -2,6 +2,7 @@ package com.loopers.application.user;
 
 import com.loopers.domain.user.UserModel;
 import com.loopers.domain.user.UserService;
+import com.loopers.interfaces.api.User.Gender;
 import com.loopers.support.error.CoreException;
 import com.loopers.support.error.ErrorType;
 import lombok.RequiredArgsConstructor;
@@ -12,17 +13,17 @@ import org.springframework.stereotype.Component;
 public class UserFacade {
     private final UserService userService;
 
-    public UserVo.UserInfo signIn(UserVo.UserCommand userCommand){
+    public UserInfo signIn(String loginId, Gender gender, String brith, String email){
 
-        UserModel userModel = userService.save(userCommand.toModel());
-        return UserVo.UserInfo.from(userModel);
+        UserModel userModel = userService.save(loginId, gender, brith, email);
+        return UserInfo.from(userModel);
     }
 
-    public UserVo.UserInfo getByLoginId(String loginId){
+    public UserInfo getByLoginId(String loginId){
         UserModel userModel = userService.getByLoginId(loginId);
         if(userModel == null){
             throw new CoreException(ErrorType.BAD_REQUEST, "존재하지 않는 사용자 입니다.");
         }
-        return UserVo.UserInfo.from(userModel);
+        return UserInfo.from(userModel);
     }
 }
