+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/User/UserV1Controller.java
@@ -1,7 +1,7 @@
 package com.loopers.interfaces.api.User;
 
 import com.loopers.application.user.UserFacade;
-import com.loopers.application.user.UserVo;
+import com.loopers.application.user.UserInfo;
 import com.loopers.interfaces.api.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -16,7 +16,13 @@ public class UserV1Controller implements UserV1ApiSpec {
     @Override
     public ApiResponse<UserV1Dto.UserResponse> SignIn(@RequestBody UserV1Dto.SignInRequest signInRequest) {
 
-        UserVo.UserInfo userInfo = userFacade.signIn(UserVo.UserCommand.from(signInRequest));
+        UserInfo userInfo = userFacade.signIn(
+                signInRequest.loginId(),
+                Gender.fromCode(signInRequest.gender()),
+                signInRequest.brith(),
+                signInRequest.email()
+        );
+
         UserV1Dto.UserResponse userResponse = UserV1Dto.UserResponse.from(userInfo);
 
         return ApiResponse.success(userResponse);
@@ -26,7 +32,7 @@ public class UserV1Controller implements UserV1ApiSpec {
     @GetMapping("/me")
     public ApiResponse<UserV1Dto.UserResponse> getByLoginId(String loginId) {
 
-        UserVo.UserInfo userInfo = userFacade.getByLoginId(loginId);
+        UserInfo userInfo = userFacade.getByLoginId(loginId);
         UserV1Dto.UserResponse userResponse = UserV1Dto.UserResponse.from(userInfo);
 
         return ApiResponse.success(userResponse);
